requirements.txt
# pandas
mlflow==2.20.1
notebook
numpy
pytorch
# scikit-learn
matplotlib
python-box==6.0.2
pyYAML==6.0.2
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
loguru
# kaggle
python-dotenv==1.1.0
pandera
pandera[io]
fastapi #==0.115.0
uvicorn #==0.31.0
pydantic #==2.9.0
boto3 #==1.35.0
sqlalchemy #==2.0.35
faiss-cpu==1.10.0
langchain-community
langchain-text-splitters
langchain-huggingface






-e . # This is needed for creation as a local Project

# jwk Notes
# python-box: subclass of dict that overrides some base functionality so dcit times can be accessed as attriure or key value
# joblib: facilitates pipelines to run in parallel
# notebook serves jupyter notebook through web interface
#  -e . : project installs by pip in an editable mode (dot referes to current directory as pkg source)
# flask: lightweight web framework
# flask-cors an extension for flask tht handles CORS (cross-origin resource sharing); required when frontend tires to access a backend api
#   hosted on a different domain
# ensure: for writing assertion helpers in a supposedly more readable form: ensure(5).is_a(int).is_positive()
# pandera for more in-depth schema use
